Added Method to get Group Posts
Added method in query_methods.rb to grab the posts from a given
group. Linkedin limits the number of posts returned to 50 so it
might be a good idea to add the logic to get all posts into this
function, and return everything as 1 big mash. Also, there is no
test coverage because I don't really know how vcr works yet.